+++ src/core_modules/capture-core/components/DataEntries/EnrollmentRegistrationEntry/hooks/useFormValues.js
@@ -148,7 +148,7 @@ export const useFormValues = ({ program, trackedEntityInstanceAttributes, orgUni
 
     useEffect(() => {
         if (
-            orgUnit?.code &&
+            orgUnit?.id &&
             Object.entries(formFoundation).length > 0 &&
             formValuesReadyRef.current === false &&
             areAttributesWithSubvaluesReady